requirements.txt
konlpy==0.5.1
hgtk==0.1.3
torchaudio==0.3.0
